pyproject.toml
[tool.poetry]
name = "goit-pythonweb-hw-08"
version = "0.1.0"
description = ""
authors = ["Alex Goliansky <alx5@ukr.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12.8"
fastapi = {extras = ["standard"], version = "^0.115.6"}
asyncpg = "^0.30.0"
sqlalchemy = "^2.0.37"
alembic = "^1.14.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
